remove v0.1 and v0.2 of Case API
